Move string simplification to helper
